+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ### Views
 
 - Multiput now supports softdeleting related models. (#24)
+
+### Router
+
+- Router is no longer a singleton (#28; **backwards incompatible**)